pyproject.toml
[project]
name = "Instring_project"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.3",
    "fastapi>=0.128.5",
    "fastapi-mail>=1.6.1",
    "psycopg2>=2.9.11",
    "pwdlib>=0.3.0",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.11.0",
    "redis>=7.1.0",
    "sqlmodel>=0.0.32",
]
